refactor: index.css파일 분할, Section파일들 Section폴더로 이동
