+++ src/web/css/asset.css
@@ -1,0 +1,21 @@
+.error-message {
+    color: #ff0000;
+    margin: 0.3rem;
+    height: 1rem;
+    width: 90%;
+}
+
+
+.button {
+    border-radius: 0.25rem;
+    background: var(--lotto-primary, #4E5BA6);
+    color: var(--lotto-greyscale-1, #FFF);
+    font-feature-settings: 'clig' off, 'liga' off;
+    height: 2rem;
+    font-family: Roboto;
+    font-size: 0.875rem;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 1rem;
+    letter-spacing: 0.07813rem;
+}
