+++ src/web/css/purchaseSection.css
@@ -1,0 +1,31 @@
+.puchase-form {
+  width: 100%;
+}
+
+.flex-container {
+  display: flex;
+}
+
+.purchase-input-text {
+  width: 20.875rem;
+  height: 1.8rem;
+  border-radius: 0.25rem;
+  border: 1px solid var(--lotto-greyscale-3, #B4B4B4);
+  margin: 0.4rem 0.4rem 0 0;
+}
+
+.purchase-button {
+  margin: 0.4rem 0 0.4rem 0;
+  flex-grow: 1;
+}
+
+.purchase-context {
+  color: var(--lotto-greyscale-5, #000);
+  font-feature-settings: 'clig' off, 'liga' off;
+  font-family: Roboto;
+  font-size: 0.9375rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 1.5rem;
+  letter-spacing: 0.03125rem;
+}
